// src/facebook-page-post/entities/facebook-insight.interface.ts
export interface IFacebookInsightData {
	value: any;
	end_time: string;
}

export interface IFacebookInsight {
	name: FacebookInsightMetric;
	period: FacebookInsightPeriod;
	values: IFacebookInsightData[];
	title?: any;
	description: string;
	id: string;
}

export interface IFacebookInsightPaging {
	previous: string;
	next: string;
}

export interface IFacebookInsightResponse {
	data: IFacebookInsight[];
	paging: IFacebookInsightPaging;
}
// ...
export enum FacebookInsightPeriod {
  DAY = "day",
  WEEK = "week",
  DAYS_28 = "days_28",
  MONTH = "month",
  LIFETIME = "lifetime",
  TOTAL_OVER_RANGE = "total_over_range",
}